requirements.txt
behave==1.2.6
python-dotenv==0.14.0
selenium==3.141.0
webdriver-manager==3.2.2